support wasm source maps
