+++ .gitignore
@@ -7,8 +7,8 @@ pubspec.lock
 # Common build output
 /*/build/
 
-# Include .packages files from tests which are hand coded
-!build_runner_core/test/fixtures/**/.dart_tool/
+# Include packages config files from tests which are hand coded
+!build_runner_core/test/fixtures/**/.dart_tool/package_config.json
 !build_runner_core/test/fixtures/**/pubspec.lock
 
 # Extra files from dart2js that we don't want
