+++ build_web_compilers/lib/src/common.dart
@@ -2,6 +2,7 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
+import 'dart:convert';
 import 'dart:io';
 
 import 'package:build/build.dart';
@@ -38,15 +39,16 @@ void validateOptions(Map<String, dynamic> config, List<String> supportedOptions,
   }
 }
 
-/// Fixes up the [uris] from a source map so they make sense in a browser
-/// context.
+/// If [id] exists, assume it is a source map and fixe up the source uris from
+/// it so they make sense in a browser context, then write the modified version
+/// using [writer].
 ///
 /// - Strips the scheme from the uri
 /// - Strips the top level directory if its not `packages`
-///
-/// Copied to `web/stack_trace_mapper.dart`, these need to be kept in sync.
-List<String> fixSourceMapSources(List<String> uris) {
-  return uris.map((source) {
+Future<void> fixAndCopySourceMap(
+    AssetId id, ScratchSpace scratchSpace, AssetWriter writer) async {
+  /// Copied to `web/stack_trace_mapper.dart`, these need to be kept in sync.
+  String fixMappedSource(String source) {
     var uri = Uri.parse(source);
     // We only want to rewrite multi-root scheme uris.
     if (uri.scheme.isEmpty) return source;
@@ -54,5 +56,17 @@ List<String> fixSourceMapSources(List<String> uris) {
         ? uri.pathSegments
         : uri.pathSegments.skip(1);
     return Uri(path: p.url.joinAll(['/', ...newSegments])).toString();
-  }).toList();
+  }
+
+  var file = scratchSpace.fileFor(id);
+  if (await file.exists()) {
+    var content = await file.readAsString();
+    var json = jsonDecode(content) as Map<String, Object?>;
+    var sources = json['sources'] as List<Object?>;
+    // Modify `sources` in place for fewer allocations.
+    for (var i = 0; i < sources.length; i++) {
+      sources[i] = fixMappedSource(sources[i] as String);
+    }
+    await writer.writeAsString(id, jsonEncode(json));
+  }
 }
